Support JDK9-native ALPN
Motivation:
Netty is unable to take advantage of JDK9's built-in ALPN support.

Modifications:
When running on JDK9 or later, Netty will invoke Java's built-in methods
for specifying the supported (server) or requested (client) application
layer protocols and for obtaining the negotiated protocol after the
handshake has taken place.

The relevant methods are new in Java 9, so they are called using core
reflection. As all of the methods in question are public, there are no
issues involving Jigsaw encapsulation or even performance (as all the
Method instances are cached in static final fields). Since it is
possible to configure a SecurityManager to prevent the use of
Class#getMethod even with public methods, AccessController#doPrivileged
is used to obtain the necessary Method instances.

Result:
Fixes #6933.