Collapse JdkApplicationProtocolNegotiator class hierarchy
Motivation:
The failIfNoCommonProtocols behavior is needlessly difficult to
implement on JDK9: SSLEngine is configured with ALPN behavior directly
instead of using ProtocolSelector/ProtocolSelectionListener strategies,
but the relevant code only has access to those (unused) strategies and
not the original config value.

Modifications:
The JdkApplicationProtocolNegotiator class hierarchy has been simplified
by collapsing it into a single class. The ProtocolSelector and
ProtocolSelectionListener interfaces have been replaced with the
EngineConfigurator class, which is not used on JDK9.

Result:
The ALPN implementation is significantly easier to understand. The
Java9SslUtils#configureAlpn method can now be passed the
failIfNoCommonProtocols flag.